agent/requirements.txt
psutil==5.9.5
pywin32==306
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0 